+++ app/src/main/java/riza/com/cto/view/check/CheckActivity.kt
@@ -90,7 +90,7 @@ class CheckActivity : AppCompatActivity(), OnMapReadyCallback, SeekBar.OnSeekBar
 
         vm.displayRadius.observe(this, Observer {
             mCircle?.remove()
-            createCircle(vm.centroid.value!!, it)
+            createCircle(vm.centroid.value!!, it.toDouble())
         })
 
         vm.listTest.observe(this, Observer {
